+++ scripts/macros/2014/classFeatures/sorcerer/aberrantMind/psychicDefenses.js
@@ -1,0 +1,5 @@
+export let psychicDefenses = {
+    name: 'Psychic Defenses',
+    version: '1.2.17',
+    rules: 'legacy'
+};